backend/requirements.txt
django-filter==25.1
django==5.1.3
djangorestframework==3.14
django-cors-headers==4.3.1
gunicorn==23.0.0